% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{cv_eif_RCT}
\alias{cv_eif_RCT}
\title{Title}
\usage{
cv_eif_RCT(
  fold,
  data_in,
  contrast,
  g_learners,
  h_learners,
  b_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  d_learners,
  effect_type = c("p_all", "p_i", "p_hat", "Y_1"),
  w_names,
  z_names,
  g_bounds = c(0.01, 0.99)
)
}
\arguments{
\item{fold}{.}

\item{data_in}{.}

\item{contrast}{.}

\item{g_learners}{.}

\item{h_learners}{.}

\item{b_learners}{.}

\item{q_learners}{.}

\item{r_learners}{.}

\item{u_learners}{.}

\item{v_learners}{.}

\item{d_learners}{.}

\item{effect_type}{.}

\item{w_names}{.}

\item{z_names}{.}

\item{g_bounds}{.}
}
\description{
Title
}
