% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{est_onestep_RCT}
\alias{est_onestep_RCT}
\title{One-step estimator}
\usage{
est_onestep_RCT(
  data,
  contrast,
  g_learners,
  h_learners,
  b_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  d_learners,
  l_learners,
  w_names,
  z_names,
  l_names,
  y_bounds,
  g_bounds = c(0.005, 0.995),
  effect_type = c("shift_k", "shift_k_order", "Y_1"),
  svy_weights = NULL,
  cv_folds = 5L,
  cv_strat = FALSE,
  strat_pmin = 0.1
)
}
\arguments{
\item{data}{A data.table containing all variables}

\item{contrast}{Numeric vector of length 2 giving treatment contrast (e.g. c(1,0)).}

\item{g_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating nuisance parameters g}

\item{h_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.}

\item{b_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating outcome regression b()}

\item{q_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating nuisance parameters q}

\item{r_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating nuisance parameters r}

\item{u_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating nuisance parameters u}

\item{v_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.
This learner is used for estimating nuisance parameters v}

\item{d_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.}

\item{l_learners}{A learner object. It is either an \code{\link[sl3]{Stack}} or any class inheriting from \code{\link[sl3]{Lrnr_base}} that contains one or more instantiated \pkg{sl3} learners.}

\item{w_names}{Character vector of column names for W}

\item{z_names}{Character vector of column names for Z}

\item{l_names}{Character vector of column names for L}

\item{y_bounds}{A numeric vector specifying the upper and lower bound for the outcome Y.}

\item{g_bounds}{A numeric vector specifying the upper and lower bound for the estimated propensity scores.}

\item{effect_type}{Character string specifying the estimand to compute.}

\item{svy_weights}{.}

\item{cv_folds}{Value specifying the number of folds to be created for cross-validation.}

\item{cv_strat}{Logical, indicating whether to perform stratified cross-fitting.}

\item{strat_pmin}{Stratified cross-fitting will be used if the outcome prevalence is below this threshold}
}
\description{
One-step estimator
}
