% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_moc_mech_RCT}
\alias{fit_moc_mech_RCT}
\title{Title fit_moc_mech_RCT}
\usage{
fit_moc_mech_RCT(
  train_data,
  valid_data = NULL,
  contrast,
  learners,
  z_names,
  w_names,
  type = c("q", "r")
)
}
\arguments{
\item{train_data}{.}

\item{valid_data}{.}

\item{contrast}{.}

\item{learners}{.}

\item{z_names}{.}

\item{w_names}{.}

\item{type}{.}
}
\description{
Title fit_moc_mech_RCT
}
