% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_nuisance_l_RCT}
\alias{fit_nuisance_l_RCT}
\title{Fitting models for l}
\usage{
fit_nuisance_l_RCT(
  train_data,
  valid_data,
  contrast,
  learners,
  b_out,
  z_names,
  w_names,
  effect_type
)
}
\arguments{
\item{train_data}{A data.table containing the training data}

\item{valid_data}{A data.table containing the validation data}

\item{contrast}{Numeric vector of length 2 giving treatment contrast (e.g. c(1,0)).}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}}

\item{b_out}{Results from outcome regression}

\item{z_names}{A character vector of the names of the columns of Z}

\item{w_names}{A character vector of the names of the columns of W}

\item{effect_type}{Specifying the estimand}
}
\description{
Fitting models for l
}
