% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_nuisance_u_RCT}
\alias{fit_nuisance_u_RCT}
\title{Fitting models for u}
\usage{
fit_nuisance_u_RCT(
  train_data,
  valid_data,
  learners,
  b_out,
  q_out,
  r_out,
  g_out,
  h_out,
  w_names
)
}
\arguments{
\item{train_data}{A data.table containing the training data}

\item{valid_data}{A data.table containing the validation data}

\item{learners}{\code{\link[sl3]{Stack}}, or other learner class (inheriting
from \code{\link[sl3]{Lrnr_base}}), containing a set of learners from
\pkg{sl3}}

\item{b_out}{Results from outcome regression}

\item{q_out}{Results from regression model for estimating q}

\item{r_out}{Results from regression model for estimating r}

\item{g_out}{Results from regression model for estimating g}

\item{h_out}{Results from regression model for estimating h}

\item{w_names}{A character vector of the names of the columns of W}
}
\description{
Fitting models for u
}
