% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_treat_mech_RCT}
\alias{fit_treat_mech_RCT}
\title{Title}
\usage{
fit_treat_mech_RCT(
  train_data,
  valid_data = NULL,
  contrast,
  learners,
  w_names,
  type = c("g", "h"),
  bounds = c(0.01, 0.99)
)
}
\arguments{
\item{train_data}{.}

\item{valid_data}{.}

\item{contrast}{.}

\item{learners}{.}

\item{w_names}{.}

\item{type}{.}

\item{bounds}{.}
}
\description{
Title
}
