% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.medoutcon}
\alias{print.medoutcon}
\title{Print method for natural/interventional (in)direct effect estimate objects (This function is directly from the medoutcon package)}
\usage{
\method{print}{medoutcon}(x, ...)
}
\arguments{
\item{x}{An object of class \code{medoutcon}.}

\item{...}{Other options (not currently used).}
}
\description{
The \code{print} method for objects of class \code{medoutcon}.
}
\keyword{internal}
