% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_to_unit}
\alias{scale_to_unit}
\title{Scale values to the unit interval (This function is directly from the medoutcon package)}
\usage{
scale_to_unit(vals)
}
\arguments{
\item{vals}{A \code{numeric} vector of values to be scaled into the closed
interval \code{0, 1}.}
}
\description{
Scale values to the unit interval (This function is directly from the medoutcon package)
}
\keyword{internal}
