% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_fun.R
\name{est_onestep_RCT}
\alias{est_onestep_RCT}
\title{Title}
\usage{
est_onestep_RCT(
  data,
  contrast,
  g_learners,
  h_learners,
  b_learners,
  q_learners,
  r_learners,
  u_learners,
  v_learners,
  d_learners,
  w_names,
  z_names,
  y_bounds,
  g_bounds = c(0.01, 0.99),
  effect_type = c("p_all", "p_i", "p_hat", "Y_1"),
  svy_weights = NULL,
  cv_folds = 5L
)
}
\arguments{
\item{data}{.}

\item{contrast}{.}

\item{g_learners}{.}

\item{h_learners}{.}

\item{b_learners}{.}

\item{q_learners}{.}

\item{r_learners}{.}

\item{u_learners}{.}

\item{v_learners}{.}

\item{d_learners}{.}

\item{w_names}{.}

\item{z_names}{.}

\item{y_bounds}{.}

\item{g_bounds}{.}

\item{effect_type}{.}

\item{svy_weights}{.}

\item{cv_folds}{.}
}
\description{
Title
}
