% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medoutconRCT.R
\name{medoutconRCT}
\alias{medoutconRCT}
\title{Title .}
\usage{
medoutconRCT(
  W,
  A,
  Z,
  M,
  Y,
  R = rep(1, length(Y)),
  obs_weights = rep(1, length(Y)),
  svy_weights = NULL,
  two_phase_weights = rep(1, length(Y)),
  contrast = c(1, 0),
  g_learners = sl3::Lrnr_glm_fast$new(),
  h_learners = sl3::Lrnr_glm_fast$new(),
  b_learners = sl3::Lrnr_glm_fast$new(),
  q_learners = sl3::Lrnr_glm_fast$new(),
  r_learners = sl3::Lrnr_glm_fast$new(),
  u_learners = sl3::Lrnr_hal9001$new(),
  v_learners = sl3::Lrnr_hal9001$new(),
  d_learners = sl3::Lrnr_glm_fast$new(),
  estimator = c("tmle", "onestep"),
  estimator_args = list(cv_folds = 5L, max_iter = 5L, tiltmod_tol = 5),
  g_bounds = c(0.01, 0.99)
)
}
\arguments{
\item{W}{.}

\item{A}{.}

\item{Z}{.}

\item{M}{.}

\item{Y}{.}

\item{R}{.}

\item{obs_weights}{.}

\item{svy_weights}{.}

\item{two_phase_weights}{.}

\item{contrast}{.}

\item{g_learners}{.}

\item{h_learners}{.}

\item{b_learners}{.}

\item{q_learners}{.}

\item{r_learners}{.}

\item{u_learners}{.}

\item{v_learners}{.}

\item{d_learners}{.}

\item{estimator}{.}

\item{estimator_args}{.}

\item{g_bounds}{.}
}
\description{
Title .
}
